New follow approach
This is a painfully un-atomic commit, but I didn't see any better way to
split it up.  It implements a new approach to following, and makes the
event emissions more consistent.

Rather than listen with include_docs=true, it omits the docs, and just
fetches each one when necessary.  This allows it to get the docs with
?revs=true&att_encoding_info=true, every time.

It would be nice to be able to specify include_attachments on the doc
fetch, but that would result in a lot of unnecessary bandwidth for
attachments already in Manta.

Also, all events now emit a 'change' object, rather than emitting the
'doc' on some, and the 'change' on others.  The 'change.doc' member is
set with the doc once fetched, with the revs and att_encoding_info
attached.  This is *not* done for deletions and non-uri-safe docs,
though, so considerable bandwidth can be saved with large design docs.